feat: convert cdk type to radio filter
